0.0.7: `LeviCivitaReal(1) + 0j` is now complex; add `calc.py`
`calc.py` runs a calculator--just a custom Python interpreter with
everything pre-imported (using real), and with `__repr__` hooked to
`__str__`.

All of the binary operators can now coerce a `LeviCivitaReal` and a
`complex` to give a `LeviCivitaComplex` result whenever `float` and
`complex` give a `complex` result (and raise a `TypeError` whenever
they do, e.g., `//` and `%`).

THis means that, even if you import `real`, you can create complex
infinitesimals like `ε*1j`, which is especially handy for the
calculator.